% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib.R
\name{get_mets}
\alias{get_mets}
\title{Compute metrics for fuels}
\usage{
get_mets(fuels, info = NULL, parallel = F, make.cluster = F)
}
\arguments{
\item{fuels}{fuel object outputs from gen_fuels()}

\item{info}{metrics info from get_mets_info(), if NULL, default metrics are used}

\item{parallel}{compute metrics in parallel}

\item{make.cluster}{make parallel cluster in function}
}
\description{
Compute metrics for each fuel realization in fuels object
}
