% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib.R
\name{mcmc}
\alias{mcmc}
\title{MCMC with joint proposal}
\usage{
mcmc(
  y_obs,
  fuel,
  prior,
  filename,
  n.samples = 10000,
  n.burn = 1000,
  stop.update = 1000,
  gen_reps = 1,
  avg = "llh",
  mets_parallel = T,
  update.every = 1,
  save.every = 1000,
  load.theta = 0,
  make.cluster = T,
  parallel = T,
  verbose = 0
)
}
\arguments{
\item{filename}{File where mcmc results should be saved. Also represents file that is loaded if load.theta=T}

\item{n.samples}{total number of mcmc samples}

\item{n.burn}{number of samples to burn and use for addapting proposal covariance}

\item{update.every}{how often to update proposal covariance during burn in phase}

\item{load.theta}{load and add to previous mcmc samples}

\item{verbose}{frequency with which status updates are printed}
}
\description{
Addaptive MCMC as defined in Haario et al. 2001 - "An adaptive Metropolis algorithm"
}
\details{
Returns predictions at X.pred.orig
}
\examples{
# See examples folder for R markdown notebooks.

}
