% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib.R
\name{mcmc_MH_adaptive}
\alias{mcmc_MH_adaptive}
\title{MCMC Calibration for FuelGen model}
\usage{
mcmc_MH_adaptive(
  y_obs,
  fuel,
  prior,
  filename = NULL,
  n.samples = 10000,
  n.burn = 1000,
  adapt.par = c(100, 20, 0.5, 0.75),
  prop.sigma = diag(0.1^2 * prior$theta_est),
  GP.init.size = 32,
  I.transform = "exp",
  gen_reps = 25,
  ABC = F,
  ABC_eps = 15,
  ABC_wt = c(1, 0.1),
  gen_parallel = F,
  mets_parallel = T,
  load.theta = 0,
  make.cluster = T,
  parallel = T,
  verbose = T,
  fixed = NULL
)
}
\arguments{
\item{y_obs}{metrics for observed data}

\item{prior}{prior from get_prior_info()}

\item{fuelsgen}{object for observed data}
}
\description{
Calibrate parameters to observed data using addaptive metropolis
}
