% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{modify_raster}
\alias{modify_raster}
\title{Fix raster}
\usage{
modify_raster(data)
}
\arguments{
\item{data:}{raster object}
}
\description{
Change raster extent and coordinates to work with the generative model. Also scale raster values to lie in \link{-1,1}. Return a list containing the raster, coordinates, and values all in the correct format for the model.
}
