% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{plot.fuelsgen}
\alias{plot.fuelsgen}
\title{Plot fuel layouts}
\usage{
\method{plot}{fuelsgen}(
  data,
  axis = F,
  circles = T,
  text.size = 15,
  point_size = 1,
  which = NULL,
  plot.dim = NULL,
  obs = NULL,
  fill = "forestgreen"
)
}
\arguments{
\item{data:}{fuels data object returned from gen_fuels}
}
\description{
Plot fuel layouts on a square tiled figure
}
