% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{plot.fuelsgen_mcmc}
\alias{plot.fuelsgen_mcmc}
\title{Diagnostic plots for mcmc calibration}
\usage{
\method{plot}{fuelsgen_mcmc}(
  mcmc,
  burn = 0,
  type = "trace",
  plot_prior = F,
  plot_llh = F,
  truth = NULL
)
}
\arguments{
\item{mcmc:}{fuelsgen_mcmc object from fuelsgen::mcmc_MH_adaptive()}

\item{burn:}{Number of initial samples to discard in plots}

\item{type:}{'trace' or 'density'. 'trace' gives trace plots for the parameters, while 'density' plots the KDE and prior over the prior range}
}
\description{
Plot a fuelsgen_mcmc object
}
