% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{plot.fuelsgen_prior}
\alias{plot.fuelsgen_prior}
\title{Plots for calibration priors}
\usage{
\method{plot}{fuelsgen_prior}(prior)
}
\arguments{
\item{prior:}{fuelsgen_prior object from fuelsgen::get_prior_info()}
}
\description{
Plot a fuelsgen_prior object
}
