% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_pairs.R
\name{scatter_pairs}
\alias{scatter_pairs}
\title{Plot samples of MCMC posteriors}
\usage{
scatter_pairs(
  theta.list,
  prior,
  truth = NULL,
  cols = c("cornflowerblue", "darkorange", "forestgreen"),
  rho.bw = "nrd0",
  mu.bw = "nrd0",
  var.bw = "nrd0",
  lambda.bw = "nrd0",
  nu.bw = "nrd0",
  labels = c("posterior"),
  pt.max = 3000,
  pt.alpha = 0.25,
  pt.cex = 0.4,
  pt.pch = 16,
  lwd = 2
)
}
\arguments{
\item{theta.list}{list of matrices containing posterior samples}

\item{truth}{optional vector of known parameter values}

\item{cols}{colors for posterior density estimate plots}

\item{labels}{legend names for posterior density estimates}
}
\description{
Pairs plot of posterior density estimates and prior distributions
}
