% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{gen_fuels}
\alias{gen_fuels}
\title{Generate fuel layouts}
\usage{
gen_fuels(
  dimX,
  dimY,
  density = 0.1/dimX/dimY,
  radius = 1,
  sd_radius = 0.1,
  height = NULL,
  sd_height = 0,
  heterogeneity = mean(c(dimX, dimY))/3,
  intensity.scale = 1,
  sp.cov.locs = NULL,
  sp.cov.vals = NULL,
  sp.cov.scale = NULL,
  reps = 1,
  GP.init.size = 28,
  seed = NULL,
  sample.method = "direct",
  verbose = T
)
}
\arguments{
\item{dimX}{(0,Inf): The size of the domain in X direction (meters)}

\item{dimY}{(0,Inf): The size of the domain in X direction (meters)}

\item{density}{(0,Inf): The mean number of fuels items per unit area (square meter)}

\item{radius}{(0,Inf): The Average radius of a circular fuel element}

\item{sd_radius}{(0,Inf): The standard deviation of fuel element radii}

\item{height}{(0,Inf): The Average height of a fuel element. Default is NULL indicating we should not sample heights.}

\item{sd_height}{(0,Inf): The standard deviation of fuel element heights. Default is 0 indicating no variance in heights.}

\item{heterogeneity}{(0,Inf): How heterogeneous the fuel maps will be. Larger values result in more heterogeneity.}

\item{intensity.scale}{(0,Inf): Standard deviation of the intensity function.}

\item{reps}{(integer): The number of fuel layouts to generate}

\item{GP.init.size}{(integer): Number of grid cells in each direction for discrete representation of the intensity function.}

\item{seed}{(integer): Optional random seed for reproducibility}

\item{verbose}{(logical): Print output related to generated fuels}

\item{intensity.mean}{(-Inf,Inf): Mean of the intensity function}

\item{parallel}{(logical): Should the fuels be generated in parallel}
}
\description{
Generate fuel layouts on a rectangular domain using a handful of parameters to a generative model.
}
