% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{gen_fuels}
\alias{gen_fuels}
\title{Generate fuel layouts}
\usage{
gen_fuels(
  dimX,
  dimY,
  density,
  heterogeneity,
  radius,
  sd_radius,
  height = NULL,
  sd_height = 0,
  GP.sd = 1,
  X.sp.locs = NULL,
  X.sp.vals = NULL,
  B = NULL,
  reps = 1,
  seed = NULL,
  subsample = T,
  subsample.size = NULL,
  batch.size = NULL,
  verbose = T
)
}
\arguments{
\item{dimX}{(0,Inf): The size of the domain in X direction (meters)}

\item{dimY}{(0,Inf): The size of the domain in X direction (meters)}

\item{density}{(0,Inf): The mean number of fuels items per unit area (square meter)}

\item{heterogeneity}{(0,Inf): How heterogeneous the fuel maps will be. Larger values result in more heterogeneity.}

\item{radius}{(0,Inf): The Average radius of a circular fuel element}

\item{sd_radius}{(0,Inf): The standard deviation of fuel element radii}

\item{height}{(0,Inf): The Average height of a fuel element. Default is NULL indicating we should not sample heights.}

\item{sd_height}{(0,Inf): The standard deviation of fuel element heights. Default is 0 indicating no variance in heights.}

\item{reps}{(integer): The number of fuel layouts to generate}

\item{seed}{(integer): Optional random seed for reproducibility}

\item{subsample}{(logical): Use subsampling to increase Gaussian Process scalability for high density layouts}

\item{subsample.size}{(integer): Size of initial GP subsample used to predict at remaining locations}

\item{batch.size}{(integer): Batch size for GP predictions}

\item{verbose}{(logical): Print status information}
}
\description{
Generate fuel layouts on a rectangular domain using a handful of parameters to a generative model.
}
