% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib.R
\name{get_mets_info}
\alias{get_mets_info}
\title{Metrics info}
\usage{
get_mets_info(
  ncc = T,
  nholes = T,
  grid.area = T,
  moran = F,
  geary = T,
  rook = F,
  queen = T,
  perim = F,
  transect = F,
  n.trans = 0
)
}
\arguments{
\item{ncc}{}

\item{nholes}{number of holes in the binary map}

\item{grid.area}{}

\item{moran}{Moran's I spatial autocorrelation. Moran=F by default because it generally doesn't add extra information from Geary.}

\item{geary}{Geary's C spacial autocorrelation}

\item{rook}{rook method for connected components in Moran/Geary. Rook is turned off my default as it generally doesn't add extra information from queen.}

\item{queen}{queen method for connected components in Moran/Geary.}

\item{perim}{perimeter of disks. Perimeter=F by default as it's very expensive to compute.}

\item{transect}{Transect lines}

\item{n.trans}{Number of transect lines layed out in a grid}
}
\description{
Get a list of metrics indicators
}
\details{
returns a list of indicators for each metric
}
