% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib.R
\name{get_prior_info}
\alias{get_prior_info}
\title{Setup prior for MCMC}
\usage{
get_prior_info(
  fuel,
  metrics,
  est_cov_obs = T,
  est_cov_samples = 25,
  est_cov_reps = 25,
  est_rho_prior = list(prior = "gamma", params = c(1, 10/fuel$dimX)),
  hcauchy_scale = 0.1,
  GP.init.size = 64,
  sample.method = "direct",
  mets_parallel = T,
  make.cluster = F,
  seed = NULL
)
}
\arguments{
\item{fuel}{observed fuel object, output from gen_fuels()}

\item{metrics}{metrics for fuel object, output from get_mets()}

\item{est_cov_obs}{estimate covariance using only observations (not recommended for n<10)}

\item{est_cov_samples}{number of samples from priors used for augmenting observed metrics, only used if est_cov_obs=F}

\item{est_cov_reps}{number of fuel replicates to generate at each sample, only used if est_cov_obs=F}

\item{est_rho_prior}{list containing string 'prior' denoting the prior distribution and vector 'params' containing the parameters of the distribution for the lengthscale. Currently only 'unif' and 'gamma' priors are supported.}

\item{mets_parallel}{compute metrics in parallel (almost always faster)}

\item{make.cluster}{should a parallel cluster be created within the function?}

\item{seed}{random seed for reproducible fuel generation and metrics}
}
\value{
List containing necessary prior precomputing for MCMC
}
\description{
Compute necessary prior information for MCMC sampling
}
