% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib.R
\name{mcmc_MH_adaptive}
\alias{mcmc_MH_adaptive}
\title{MCMC using adaptive metropolis implemented in MHadaptive package}
\usage{
mcmc_MH_adaptive(
  y_obs,
  fuel,
  prior,
  filename = NULL,
  n.samples = 10000,
  n.burn = 1000,
  adapt.par = c(100, 20, 0.5, 0.75),
  prop.sigma = NULL,
  gen_reps = 25,
  avg = "llh",
  gen_parallel = F,
  mets_parallel = T,
  make.cluster = T
)
}
\arguments{
\item{y_obs}{Metrics for observed data}

\item{fuel}{Observed fuelsgen data object}

\item{filename}{File where mcmc results should be saved. Also represents file that is loaded if load.theta=T}

\item{n.samples}{total number of mcmc samples}

\item{n.burn}{number of samples to burn and use for addapting proposal covariance}

\item{adapt.par}{adaptive parameters for MHadaptive::MetroHastings}

\item{prop.sigma}{Initial proposal covariance. Default is diagonal with 10\% variance.}

\item{gen_reps}{Number of realizations to generate at each parameter setting. Default of 25 has worked well over a range of problems balancing computation time with accuracy of expectation.}

\item{avg}{How to compute the likelihood over realizations. 'llh' is an average likelihood over simulated realizations, and 'mets' computing a single likeklihood with metrics averaged over realizations.}

\item{gen_parallel}{Generate simulated fuels in parallel. Generally not faster except for large domains.}

\item{mets_parallel}{Generate metrics in parallel. Generally faster than sequential.}

\item{make.cluster}{Should the parallel computing cluster be created inside the function}
}
\description{
Addaptive MCMC as defined in Haario et al. 2001 - "An adaptive Metropolis algorithm"
}
\details{
Returns predictions at X.pred.orig
}
\examples{
# See examples folder for R markdown notebooks.

}
