% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{plot.fuelsgen}
\alias{plot.fuelsgen}
\title{Plot fuel layouts}
\usage{
\method{plot}{fuelsgen}(data)
}
\arguments{
\item{data:}{fuels data object returned from gen_fuels}
}
\description{
Plot fuel layouts on a square tiled figure
}
