% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{plot_fuels}
\alias{plot_fuels}
\title{Plot fuel layouts}
\usage{
plot_fuels(data)
}
\arguments{
\item{data:}{fuels data object returned from gen_fuels}
}
\description{
Plot fuel layouts on a square tiled figure
}
