% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{plot_fuels}
\alias{plot_fuels}
\title{Generate fuel layouts}
\usage{
plot_fuels(data, filename = NULL)
}
\arguments{
\item{filename}{(string): file path for pdf output}

\item{data:}{object returned from gen_fuels}
}
\description{
Generate fuel layouts on a rectangular domain using a handful of parameters to a generative model.
}
